+++ common/src/main/java/com/food/common/order/business/internal/OrderMenuCommonService.java
@@ -1,6 +1,6 @@
 package com.food.common.order.business.internal;
 
-import com.food.common.order.business.internal.dto.OrderMenuDto;
+import com.food.common.order.business.model.OrderMenuDto;
 
 import java.util.List;
